cleaned up join-stable page
